+++ src/main/java/org/opensearch/knn/index/util/Nmslib.java
@@ -73,4 +73,8 @@ class Nmslib extends NativeLibrary {
     public Float distanceToRadialThreshold(Float distance, SpaceType spaceType) {
         return distance;
     }
+
+    public Float scoreToRadialThreshold(Float score, SpaceType spaceType) {
+        return score;
+    }
 }