+++ src/test/java/org/opensearch/knn/index/FaissIT.java
@@ -278,7 +278,7 @@ public class FaissIT extends KNNRestTestCase {
     }
 
     @SneakyThrows
-    public void testEndToEnd_whenDoRadiusSearch_whenMethodIsHNSWFlat_thenSucceed() {
+    public void testEndToEnd_whenDoRadiusSearch_whenDistanceThreshold_whenMethodIsHNSWFlat_thenSucceed() {
         KNNMethod hnswMethod = KNNEngine.FAISS.getMethod(KNNConstants.METHOD_HNSW);
         SpaceType spaceType = SpaceType.L2;
 
@@ -329,17 +329,135 @@ public class FaissIT extends KNNRestTestCase {
         refreshAllNonSystemIndices();
         assertEquals(testData.indexData.docs.length, getDocCount(INDEX_NAME));
 
-        float radius = 300000000000f;
-        for (float[] queryVector : testData.queries) {
-            validateRadiusSearchResults(INDEX_NAME, FIELD_NAME, queryVector, radius, spaceType);
-        }
+        float distance = 300000000000f;
+        validateRadiusSearchResults(INDEX_NAME, FIELD_NAME, testData.queries, distance, null, spaceType);
 
         // Delete index
         deleteKNNIndex(INDEX_NAME);
     }
 
     @SneakyThrows
-    public void testEndToEnd_whenDoRadiusSearch_whenMethodIsHNSWPQ_thenSucceed() {
+    public void testEndToEnd_whenDoRadiusSearch_whenScoreThreshold_whenMethodIsHNSWFlat_thenSucceed() {
+        KNNMethod hnswMethod = KNNEngine.FAISS.getMethod(KNNConstants.METHOD_HNSW);
+        SpaceType spaceType = SpaceType.L2;
+
+        List<Integer> mValues = ImmutableList.of(16, 32, 64, 128);
+        List<Integer> efConstructionValues = ImmutableList.of(16, 32, 64, 128);
+        List<Integer> efSearchValues = ImmutableList.of(16, 32, 64, 128);
+
+        Integer dimension = testData.indexData.vectors[0].length;
+
+        // Create an index
+        XContentBuilder builder = XContentFactory.jsonBuilder()
+            .startObject()
+            .startObject("properties")
+            .startObject(FIELD_NAME)
+            .field("type", "knn_vector")
+            .field("dimension", dimension)
+            .startObject(KNNConstants.KNN_METHOD)
+            .field(KNNConstants.NAME, hnswMethod.getMethodComponent().getName())
+            .field(KNNConstants.METHOD_PARAMETER_SPACE_TYPE, spaceType.getValue())
+            .field(KNNConstants.KNN_ENGINE, KNNEngine.FAISS.getName())
+            .startObject(KNNConstants.PARAMETERS)
+            .field(KNNConstants.METHOD_PARAMETER_M, mValues.get(random().nextInt(mValues.size())))
+            .field(KNNConstants.METHOD_PARAMETER_EF_CONSTRUCTION, efConstructionValues.get(random().nextInt(efConstructionValues.size())))
+            .field(KNNConstants.METHOD_PARAMETER_EF_SEARCH, efSearchValues.get(random().nextInt(efSearchValues.size())))
+            .endObject()
+            .endObject()
+            .endObject()
+            .endObject()
+            .endObject();
+
+        Map<String, Object> mappingMap = xContentBuilderToMap(builder);
+        String mapping = builder.toString();
+
+        createKnnIndex(INDEX_NAME, mapping);
+        assertEquals(new TreeMap<>(mappingMap), new TreeMap<>(getIndexMappingAsMap(INDEX_NAME)));
+
+        // Index the test data
+        for (int i = 0; i < testData.indexData.docs.length; i++) {
+            addKnnDoc(
+                INDEX_NAME,
+                Integer.toString(testData.indexData.docs[i]),
+                FIELD_NAME,
+                Floats.asList(testData.indexData.vectors[i]).toArray()
+            );
+        }
+
+        // Assert we have the right number of documents
+        refreshAllNonSystemIndices();
+        assertEquals(testData.indexData.docs.length, getDocCount(INDEX_NAME));
+
+        float score = 0.00001f;
+
+        validateRadiusSearchResults(INDEX_NAME, FIELD_NAME, testData.queries, null, score, spaceType);
+
+        // Delete index
+        deleteKNNIndex(INDEX_NAME);
+    }
+
+    @SneakyThrows
+    public void testEndToEnd_whenDoRadiusSearch_whenMoreThanOneScoreThreshold_whenMethodIsHNSWFlat_thenSucceed() {
+        KNNMethod hnswMethod = KNNEngine.FAISS.getMethod(KNNConstants.METHOD_HNSW);
+        SpaceType spaceType = SpaceType.INNER_PRODUCT;
+
+        List<Integer> mValues = ImmutableList.of(16, 32, 64, 128);
+        List<Integer> efConstructionValues = ImmutableList.of(16, 32, 64, 128);
+        List<Integer> efSearchValues = ImmutableList.of(16, 32, 64, 128);
+
+        Integer dimension = testData.indexData.vectors[0].length;
+
+        // Create an index
+        XContentBuilder builder = XContentFactory.jsonBuilder()
+            .startObject()
+            .startObject("properties")
+            .startObject(FIELD_NAME)
+            .field("type", "knn_vector")
+            .field("dimension", dimension)
+            .startObject(KNNConstants.KNN_METHOD)
+            .field(KNNConstants.NAME, hnswMethod.getMethodComponent().getName())
+            .field(KNNConstants.METHOD_PARAMETER_SPACE_TYPE, spaceType.getValue())
+            .field(KNNConstants.KNN_ENGINE, KNNEngine.FAISS.getName())
+            .startObject(KNNConstants.PARAMETERS)
+            .field(KNNConstants.METHOD_PARAMETER_M, mValues.get(random().nextInt(mValues.size())))
+            .field(KNNConstants.METHOD_PARAMETER_EF_CONSTRUCTION, efConstructionValues.get(random().nextInt(efConstructionValues.size())))
+            .field(KNNConstants.METHOD_PARAMETER_EF_SEARCH, efSearchValues.get(random().nextInt(efSearchValues.size())))
+            .endObject()
+            .endObject()
+            .endObject()
+            .endObject()
+            .endObject();
+
+        Map<String, Object> mappingMap = xContentBuilderToMap(builder);
+        String mapping = builder.toString();
+
+        createKnnIndex(INDEX_NAME, mapping);
+        assertEquals(new TreeMap<>(mappingMap), new TreeMap<>(getIndexMappingAsMap(INDEX_NAME)));
+
+        // Index the test data
+        for (int i = 0; i < testData.indexData.docs.length; i++) {
+            addKnnDoc(
+                INDEX_NAME,
+                Integer.toString(testData.indexData.docs[i]),
+                FIELD_NAME,
+                Floats.asList(testData.indexData.vectors[i]).toArray()
+            );
+        }
+
+        // Assert we have the right number of documents
+        refreshAllNonSystemIndices();
+        assertEquals(testData.indexData.docs.length, getDocCount(INDEX_NAME));
+
+        float score = 5f;
+
+        validateRadiusSearchResults(INDEX_NAME, FIELD_NAME, testData.queries, null, score, spaceType);
+
+        // Delete index
+        deleteKNNIndex(INDEX_NAME);
+    }
+
+    @SneakyThrows
+    public void testEndToEnd_whenDoRadiusSearch__whenDistanceThreshold_whenMethodIsHNSWPQ_thenSucceed() {
         String indexName = "test-index";
         String fieldName = "test-field";
         String trainingIndexName = "training-index";
@@ -415,11 +533,9 @@ public class FaissIT extends KNNRestTestCase {
         refreshAllNonSystemIndices();
         assertEquals(testData.indexData.docs.length, getDocCount(indexName));
 
-        float radius = 300000000000f;
+        float distance = 300000000000f;
 
-        for (float[] queryVector : testData.queries) {
-            validateRadiusSearchResults(indexName, fieldName, queryVector, radius, spaceType);
-        }
+        validateRadiusSearchResults(indexName, fieldName, testData.queries, distance, null, spaceType);
 
         // Delete index
         deleteKNNIndex(indexName);
@@ -1575,30 +1691,40 @@ public class FaissIT extends KNNRestTestCase {
     private void validateRadiusSearchResults(
         String indexName,
         String fieldName,
-        float[] queryVector,
-        float radius,
+        float[][] queryVectors,
+        Float distanceThreshold,
+        Float scoreThreshold,
         final SpaceType spaceType
     ) throws IOException, ParseException {
-        XContentBuilder queryBuilder = XContentFactory.jsonBuilder().startObject().startObject("query");
-        queryBuilder.startObject("knn");
-        queryBuilder.startObject(fieldName);
-        queryBuilder.field("vector", queryVector);
-        queryBuilder.field("distance", radius);
-        queryBuilder.endObject();
-        queryBuilder.endObject();
-        queryBuilder.endObject().endObject();
-        final String responseBody = EntityUtils.toString(searchKNNIndex(indexName, queryBuilder, 10).getEntity());
-
-        List<KNNResult> knnResults = parseSearchResponse(responseBody, fieldName);
-
-        for (KNNResult knnResult : knnResults) {
-            float[] vector = Floats.toArray(Arrays.stream(knnResult.getVector()).collect(Collectors.toList()));
-            if (spaceType == SpaceType.L2) {
-                assertTrue(KNNScoringUtil.l2Squared(queryVector, vector) <= radius);
-            } else if (spaceType == SpaceType.INNER_PRODUCT) {
-                assertTrue(KNNScoringUtil.innerProduct(queryVector, vector) >= radius);
+        for (float[] queryVector : queryVectors) {
+            XContentBuilder queryBuilder = XContentFactory.jsonBuilder().startObject().startObject("query");
+            queryBuilder.startObject("knn");
+            queryBuilder.startObject(fieldName);
+            queryBuilder.field("vector", queryVector);
+            if (distanceThreshold != null) {
+                queryBuilder.field("distance", distanceThreshold);
+            } else if (scoreThreshold != null) {
+                queryBuilder.field("score", scoreThreshold);
             } else {
-                throw new IllegalArgumentException("Invalid space type");
+                throw new IllegalArgumentException("Invalid threshold");
+            }
+            queryBuilder.endObject();
+            queryBuilder.endObject();
+            queryBuilder.endObject().endObject();
+            final String responseBody = EntityUtils.toString(searchKNNIndex(indexName, queryBuilder, 10).getEntity());
+
+            List<KNNResult> knnResults = parseSearchResponse(responseBody, fieldName);
+
+            for (KNNResult knnResult : knnResults) {
+                float[] vector = knnResult.getVector();
+                float distance = TestUtils.computeDistFromSpaceType(spaceType, vector, queryVector);
+                if (spaceType == SpaceType.L2) {
+                    assertTrue(KNNScoringUtil.l2Squared(queryVector, vector) <= distance);
+                } else if (spaceType == SpaceType.INNER_PRODUCT) {
+                    assertTrue(KNNScoringUtil.innerProduct(queryVector, vector) >= distance);
+                } else {
+                    throw new IllegalArgumentException("Invalid space type");
+                }
             }
         }
     }