+++ src/test/java/org/opensearch/knn/index/util/AbstractKNNLibraryTests.java
@@ -132,6 +132,10 @@ public class AbstractKNNLibraryTests extends KNNTestCase {
             return 0f;
         }
 
+        public Float scoreToRadialThreshold(Float score, SpaceType spaceType) {
+            return 0f;
+        }
+
         @Override
         public int estimateOverheadInKB(KNNMethodContext knnMethodContext, int dimension) {
             return 0;