+++ src/test/java/org/opensearch/knn/index/util/NativeLibraryTests.java
@@ -69,5 +69,10 @@ public class NativeLibraryTests extends KNNTestCase {
         public Float distanceToRadialThreshold(Float distance, SpaceType spaceType) {
             return 0.0f;
         }
+
+        @Override
+        public Float scoreToRadialThreshold(Float score, SpaceType spaceType) {
+            return 0.0f;
+        }
     }
 }